day14/git-workflows-system/backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
python-multipart==0.0.6
aiofiles==23.2.1
httpx==0.25.2
python-dotenv==1.0.0
pytest==7.4.3
pytest-asyncio==0.21.1
GitPython==3.1.40
PyYAML==6.0.1
cryptography==42.0.8
requests==2.31.0
websockets==12.0
